refactor[service]: move test service implementations to impl package
