+++ src/test/java/com/OnlineShop/service/impl/CategoryServiceTest.java
@@ -1,10 +1,11 @@
-package com.OnlineShop.service;
+package com.OnlineShop.service.impl;
 
 import com.OnlineShop.entity.Category;
 import com.OnlineShop.exception.AlreadyExistsException;
 import com.OnlineShop.exception.NotFoundException;
 import com.OnlineShop.repository.ICategoryRepository;
 
+import com.OnlineShop.service.ICategoryService;
 import com.OnlineShop.service.impl.CategoryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
