+++ src/test/java/com/OnlineShop/service/impl/RoleServiceTest.java
@@ -1,10 +1,11 @@
-package com.OnlineShop.service;
+package com.OnlineShop.service.impl;
 
 import com.OnlineShop.entity.AppRole;
 import com.OnlineShop.enums.RoleEnum;
 import com.OnlineShop.exception.AlreadyExistsException;
 import com.OnlineShop.exception.NotFoundException;
 import com.OnlineShop.repository.IRoleRepository;
+import com.OnlineShop.service.IRoleService;
 import com.OnlineShop.service.impl.RoleService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
